Покрытие 91 процентов блока функций + перенос их их main
